+++ src/main/java/ru/job4j/social/controller/user/UserController.java
@@ -1,9 +1,9 @@
 package ru.job4j.social.controller.user;
 
 import lombok.AllArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import ru.job4j.social.model.User;
 import ru.job4j.social.service.user.UserService;
 
@@ -15,26 +15,37 @@ public class UserController {
     private final UserService userService;
 
     @GetMapping("/{userId}")
-    public ResponseEntity<User> get(@PathVariable("userId") Long userId) {
-        return userService.findById(userId)
+    public ResponseEntity<User> get(@PathVariable("userId") Long id) {
+        return userService.findById(id)
                 .map(ResponseEntity::ok)
                 .orElseGet(() -> ResponseEntity.notFound().build());
     }
 
     @PostMapping
     public ResponseEntity<User> createUser(@RequestBody User user) {
-        return ResponseEntity.ok(userService.createUser(user));
+        userService.createUser(user);
+        var uri = ServletUriComponentsBuilder
+                .fromCurrentRequest()
+                .path("/{id}")
+                .buildAndExpand(user.getId())
+                .toUri();
+        return ResponseEntity.created(uri).build();
     }
 
     @DeleteMapping("/{userId}")
-    @ResponseStatus(HttpStatus.OK)
-    public void deleteUserById(@PathVariable("userId") Long id) {
+    public ResponseEntity<Void> deleteUserById(@PathVariable("userId") Long id) {
+        if (userService.findById(id).isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
         userService.deleteUserById(id);
+        return ResponseEntity.noContent().build();
     }
 
     @PutMapping
-    @ResponseStatus(HttpStatus.OK)
-    public void updateUser(@RequestBody User user) {
-        userService.createUser(user);
+    public ResponseEntity<Void> updateUser(@RequestBody User user) {
+        if (userService.createUser(user) != null) {
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.notFound().build();
     }
 }
